# specs/003-backend-implementation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Full-Stack Todo API
  description: Backend API for the Full-Stack Todo application
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /tasks/{task_id}/tags:
    post:
      tags: [Tasks]
      summary: Add tags to task
      operationId: addTaskTags
      parameters:
        - $ref: '#/components/parameters/taskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_ids]
              properties:
                tag_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Tags added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: Remove tags from task
      operationId: removeTaskTags
      parameters:
        - $ref: '#/components/parameters/taskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_ids]
              properties:
                tag_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Tags removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Projects
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Projects]
      summary: Create project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get project by ID
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/projectIdParam'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Projects]
      summary: Update project
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/projectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Projects]
      summary: Delete project
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/projectIdParam'
      responses:
        '204':
          description: Project deleted (tasks moved to inbox)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Tags
  /tags:
    get:
      tags: [Tags]
      summary: List tags
      operationId: listTags
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Tags]
      summary: Create tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /tags/{tag_id}:
    get:
      tags: [Tags]
      summary: Get tag by ID
      operationId: getTag
      parameters:
        - $ref: '#/components/parameters/tagIdParam'
      responses:
        '200':
          description: Tag details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Tags]
      summary: Update tag
      operationId: updateTag
      parameters:
        - $ref: '#/components/parameters/tagIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdate'
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Tags]
      summary: Delete tag
      operationId: deleteTag
      parameters:
        - $ref: '#/components/parameters/tagIdParam'
      responses:
        '204':
          description: Tag deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth
  parameters:
    taskIdParam:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    projectIdParam:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    tagIdParam:
      name: tag_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  schemas:
    Priority:
      type: string
      enum: [CRITICAL, HIGH, MEDIUM, LOW]
      default: MEDIUM

    # Task Schemas
    TaskBase:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
        deadline:
          type: string
          format: date-time
          nullable: true
        time_estimate:
          type: integer
          minimum: 0
          nullable: true
          description: Estimated minutes to complete
        project_id:
          type: string
          format: uuid
          nullable: true
...
    Task:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          required: [id, title, priority, is_completed, user_id, created_at, updated_at]
          properties:
            id:
              type: string
              format: uuid
            is_completed:
              type: boolean
            user_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            is_overdue:
              type: boolean
              readOnly: true
            days_until_deadline:
              type: integer
              nullable: true
              readOnly: true
            tags:
              type: array
              items:
                $ref: '#/components/schemas/TagSummary'
...
    TaskResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Task'
...
    ProjectBase:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
    ProjectCreate:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
        - type: object
          required: [name]
    ProjectUpdate:
      $ref: '#/components/schemas/ProjectBase'
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
        - type: object
          required: [id, name, user_id, created_at, updated_at]
          properties:
            id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            task_count:
              type: integer
              readOnly: true
            completed_task_count:
              type: integer
              readOnly: true
    ProjectResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Project'
    ProjectListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Tag Schemas
    TagBase:
      type: object
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          nullable: true
          example: '#FF5733'
    TagCreate:
      allOf:
        - $ref: '#/components/schemas/TagBase'
        - type: object
          required: [label]
    TagUpdate:
      $ref: '#/components/schemas/TagBase'
    Tag:
      allOf:
        - $ref: '#/components/schemas/TagBase'
        - type: object
          required: [id, label, user_id, created_at]
          properties:
            id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            task_count:
              type: integer
              readOnly: true
    TagSummary:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        color:
          type: string
          nullable: true
    TagResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Tag'
    TagListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Common Schemas
    Pagination:
      type: object
      required: [offset, limit, total]
      properties:
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
              example: Task with ID xyz not found
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: BAD_REQUEST
              message: Invalid request body
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: RESOURCE_NOT_FOUND
              message: Resource not found
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: CONFLICT
              message: Tag with this label already exists
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: title is required
security:
  - bearerAuth: []